Be sure to check conditions in both directions.
Before, it was possible for e.g. a TOC only ticket to be returned when
the outward leg was that TOC but the return leg was another, or for
a station restriction to be missed in one direction.
Also check used operators are a subset of the restriction, not equality.